+++ pos-core/src/main/java/com/saulpos/core/hardware/scale/ScaleAdapter.java
@@ -1,0 +1,6 @@
+package com.saulpos.core.hardware.scale;
+
+public interface ScaleAdapter {
+
+    ScaleReadResult read(ScaleReadRequest request);
+}